+++ node/src/test/kotlin/net/corda/node/services/vault/VaultQueryTests.kt
@@ -241,6 +241,11 @@ abstract class VaultQueryTestsBase : VaultQueryParties {
             val criteria = VaultQueryCriteria(participants = listOf(BIG_CORP))
             val results = vaultService.queryBy<ContractState>(criteria)
             assertThat(results.states).hasSize(1)
+
+            // same query using strict participant matching
+            val strictCriteria = VaultQueryCriteria().withExactParticipants(listOf(BIG_CORP))
+            val strictResults = vaultService.queryBy<ContractState>(strictCriteria)
+            assertThat(strictResults.states).hasSize(0) // all states include node identity (MEGA_CORP)
         }
     }
 
@@ -254,6 +259,11 @@ abstract class VaultQueryTestsBase : VaultQueryParties {
             val criteria = VaultQueryCriteria(participants = listOf(MINI_CORP, BIG_CORP))
             val results = vaultService.queryBy<ContractState>(criteria)
             assertThat(results.states).hasSize(2)
+
+            // same query using strict participant matching
+            val strictCriteria = VaultQueryCriteria().withExactParticipants(listOf(MEGA_CORP, BIG_CORP))
+            val strictResults = vaultService.queryBy<ContractState>(strictCriteria)
+            assertThat(strictResults.states).hasSize(1)
         }
     }
 
@@ -795,10 +805,48 @@ abstract class VaultQueryTestsBase : VaultQueryParties {
             identitySvc.verifyAndRegisterIdentity(BIG_CORP_IDENTITY)
             vaultFiller.fillWithSomeTestLinearStates(2, "TEST", participants = listOf(MEGA_CORP, MINI_CORP))
             vaultFiller.fillWithSomeTestDeals(listOf("456"), participants = listOf(MEGA_CORP, BIG_CORP))
-            vaultFiller.fillWithSomeTestDeals(listOf("123", "789"), participants = listOf(BIG_CORP))
-            val criteria = LinearStateQueryCriteria(participants = listOf(BIG_CORP))
+            vaultFiller.fillWithSomeTestDeals(listOf("123", "789"), participants = listOf(MEGA_CORP))
+
+            val criteria = LinearStateQueryCriteria(participants = listOf(MEGA_CORP))
             val results = vaultService.queryBy<ContractState>(criteria)
-            assertThat(results.states).hasSize(3)
+            assertThat(results.states).hasSize(5)
+
+            // same query using strict participant matching
+            val strictCriteria = LinearStateQueryCriteria().withExactParticipants(listOf(MEGA_CORP))
+            val strictResults = vaultService.queryBy<ContractState>(strictCriteria)
+            assertThat(strictResults.states).hasSize(2)
+        }
+    }
+
+    @Test
+    fun `unconsumed dummy states for exact single participant`() {
+        database.transaction {
+            identitySvc.verifyAndRegisterIdentity(BIG_CORP_IDENTITY)
+            vaultFiller.fillWithDummyState(participants = listOf(MEGA_CORP, MINI_CORP))
+            vaultFiller.fillWithDummyState(participants = listOf(MEGA_CORP, BIG_CORP))
+            vaultFiller.fillWithDummyState(participants = listOf(MEGA_CORP)) // exact match
+            val strictCriteria = VaultQueryCriteria(exactParticipants = listOf(MEGA_CORP))
+            val strictResults = vaultService.queryBy<ContractState>(strictCriteria)
+            assertThat(strictResults.states).hasSize(1)
+        }
+    }
+
+    @Test
+    fun `unconsumed dummy states for exact two participants`() {
+        database.transaction {
+            identitySvc.verifyAndRegisterIdentity(BIG_CORP_IDENTITY)
+            vaultFiller.fillWithDummyState(participants = listOf(MEGA_CORP, MINI_CORP))
+            vaultFiller.fillWithDummyState(participants = listOf(MEGA_CORP, BIG_CORP))  // exact match
+            vaultFiller.fillWithDummyState(participants = listOf(MEGA_CORP))
+
+            val strictCriteria = VaultQueryCriteria(exactParticipants = listOf(MEGA_CORP, BIG_CORP))
+            val strictResults = vaultService.queryBy<ContractState>(strictCriteria)
+            assertThat(strictResults.states).hasSize(1)
+
+            // same query using strict participant matching (unordered list of participants)
+            val strictCriteriaUnordered = VaultQueryCriteria(exactParticipants = listOf(BIG_CORP, MEGA_CORP))
+            val strictResultsUnordered = vaultService.queryBy<ContractState>(strictCriteriaUnordered)
+            assertThat(strictResultsUnordered.states).hasSize(1)
         }
     }
 
@@ -813,8 +861,19 @@ abstract class VaultQueryTestsBase : VaultQueryParties {
             val criteria = LinearStateQueryCriteria(participants = listOf(BIG_CORP, MINI_CORP))
             val results = vaultService.queryBy<ContractState>(criteria)
             // DOCEND VaultQueryExample5
-
             assertThat(results.states).hasSize(3)
+
+            // same query using strict participant matching
+            // DOCSTART VaultQueryExample51
+            val strictCriteria = LinearStateQueryCriteria(exactParticipants = listOf(MEGA_CORP, BIG_CORP))
+            val strictResults = vaultService.queryBy<ContractState>(strictCriteria)
+            // DOCEND VaultQueryExample51
+            assertThat(strictResults.states).hasSize(1)
+
+            // same query using strict participant matching (unordered list of participants)
+            val strictCriteriaUnordered = LinearStateQueryCriteria(exactParticipants = listOf(BIG_CORP, MEGA_CORP))
+            val strictResultsUnordered = vaultService.queryBy<ContractState>(strictCriteriaUnordered)
+            assertThat(strictResultsUnordered.states).hasSize(1)
         }
     }
 
@@ -1959,6 +2018,13 @@ abstract class VaultQueryTestsBase : VaultQueryParties {
             // DOCEND VaultQueryExample11
 
             assertThat(results.states).hasSize(1)
+
+            // same query using strict participant matching
+            // DOCSTART VaultQueryExample52
+            val strictCriteria = LinearStateQueryCriteria().withExactParticipants(parties)
+            val strictResults = vaultService.queryBy<ContractState>(strictCriteria)
+            // DOCEND VaultQueryExample52
+            assertThat(strictResults.states).hasSize(0) // node identity included (MEGA_CORP)
         }
     }
 
@@ -2352,6 +2418,11 @@ abstract class VaultQueryTestsBase : VaultQueryParties {
 
             assertThat(results.states).hasSize(1)
             assertThat(results.states[0].state.data.linearId.externalId).isEqualTo("TEST1")
+
+            // same query using strict participant matching
+            val strictCriteria = LinearStateQueryCriteria().withExactParticipants(listOf(ALICE))
+            val strictResults = vaultService.queryBy<ContractState>(strictCriteria)
+            assertThat(strictResults.states).hasSize(0) // all states include node identity (MEGA_CORP)
         }
     }
 
@@ -2369,6 +2440,11 @@ abstract class VaultQueryTestsBase : VaultQueryParties {
 
             assertThat(results.states).hasSize(1)
             assertThat(results.states[0].state.data.linearId.externalId).isEqualTo("TEST1")
+
+            // same query using strict participant matching
+            val strictCriteria = LinearStateQueryCriteria().withExactParticipants(listOf(MEGA_CORP, ALICE, BOB, CHARLIE))
+            val strictResults = vaultService.queryBy<ContractState>(strictCriteria)
+            assertThat(strictResults.states).hasSize(1)
         }
     }
 
@@ -2401,7 +2477,7 @@ abstract class VaultQueryTestsBase : VaultQueryParties {
     }
 
     @Test
-    fun `composite query for fungible and linear states for multiple participants`() {
+    fun `composite query for fungible, linear and dummy states for multiple participants`() {
         database.transaction {
             identitySvc.verifyAndRegisterIdentity(ALICE_IDENTITY)
             identitySvc.verifyAndRegisterIdentity(BOB_IDENTITY)