+++ docs/source/changelog.rst
@@ -6,6 +6,10 @@ release, see :doc:`app-upgrade-notes`.
 
 Unreleased
 ----------
+* Moved and renamed the testing web server to the ``testing`` subproject. Also renamed the published artifact to ``corda-testserver.jar``.
+
+* New Vault Query criteria to specify exact matches for specified participants.
+
 * Support for Java 11 (compatibility mode). Please read https://github.com/corda/corda/pull/5356.
 
 * Removed the RPC exception privacy feature. Previously, in production mode, the exceptions thrown on the node were stripped of all content